+++ app/src/main/java/com/the_tinkering/wk/db/model/SubjectEntityDefinition.java
@@ -34,7 +34,7 @@ public final class SubjectEntityDefinition {
     @ColumnInfo(name = "object")
     public @Nullable SubjectType type;
     @ColumnInfo(name = "typeCode")
-    public int stars = 0;
+    public int numStars = 0;
     public @Nullable Long hiddenAt;
     public int lessonPosition = 0;
     public long srsSystemId = 0L;