+++ app/src/main/java/com/the_tinkering/wk/views/SubjectInfoView.java
@@ -80,6 +80,7 @@ public final class SubjectInfoView extends LinearLayout implements SubjectChange
     private final ViewProxy onYomi = new ViewProxy();
     private final ViewProxy kunYomi = new ViewProxy();
     private final ViewProxy nanori = new ViewProxy();
+    private final ViewProxy starsRating = new ViewProxy();
     private final ViewProxy meaningDivider = new ViewProxy();
     private final ViewProxy meaningMnemonic = new ViewProxy();
     private final ViewProxy meaningHint = new ViewProxy();
@@ -269,6 +270,7 @@ public final class SubjectInfoView extends LinearLayout implements SubjectChange
         onYomi.setDelegate(this, R.id.onYomi);
         kunYomi.setDelegate(this, R.id.kunYomi);
         nanori.setDelegate(this, R.id.nanori);
+        starsRating.setDelegate(this, R.id.starsRating);
         meaningDivider.setDelegate(this, R.id.meaningDivider);
         meaningMnemonic.setDelegate(this, R.id.meaningMnemonic);
         meaningHint.setDelegate(this, R.id.meaningHint);
@@ -529,6 +531,10 @@ public final class SubjectInfoView extends LinearLayout implements SubjectChange
         nanori.setText(subject.getNanoriRichText());
         nanori.setJapaneseLocale();
         nanori.setVisibility(showReadingAnswers && subject.hasNanori());
+
+        // Stars rating
+        starsRating.setVisibility(GlobalSettings.Other.getEnableStarsRatings());
+        starsRating.setSubject(subject);
 
         // Meaning mnemonic
         meaningDivider.setVisibility(subject.hasMeaningMnemonic() && (showMeaningRelated || showMeaningAnswers && subject.hasMeaningSynonyms()));